refactor: changed import type
