+++ src/Pages/Following/Following.jsx
@@ -1,6 +1,4 @@
-import FollowComp from "../../Components/FollowComp/FollowComp";
-import Footer from "../../Components/Footer/Footer";
-import Header from "../../Components/Header/Header";
+import { FollowComp, Footer, Header } from "../../Components";
 
 const Following = () => {
   return (
